requirements.txt
openai>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
watchdog>=3.0.0
sounddevice>=0.4.6
scipy>=1.11.0
pyperclip>=1.8.2
